경매기능 구현(배송 전)

+++ kh16Final3/src/main/java/com/kh/final3/dto/AttachmentDto.java
@@ -12,15 +12,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AttachmentDto {
-
     private Long attachmentNo;
     private String mediaType;
     private String path;
     private String category;
-    private Integer parentPkNo;
+    private Long parentPkNo;
     private String originalName;
     private String storedName;
     private Integer size;
     private LocalDateTime createdTime;
-
 }
